backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
motor==3.4.0
celery==5.4.0
redis==5.0.4
pandas==2.2.2
openpyxl==3.1.2
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.1
pydantic-settings==2.2.1
httpx==0.27.0
python-dotenv==1.0.1
